Refaktorer brevperiodelogikk
